Read and write latest versions to a file to avoid API calls.
GitHub Actions run in virtual machines that can easily be rate-
limited when using the GitHub API. There's no way to know if the
current machine has made requests recently, so this leads to flaky
failures. Instead, we'll use an action to write the latest versions
to disk and use that in subsequent runs.